1. Create a new Component called FirstCharForBrandName. it will used to  show campaign Brand name first cherecter with differen brackbrand color 2. add this re-useable components into CampaignCardForPlan and CampaignsListModel componnets
